ajuste na paginação e icones de sessão
