+++ sei/web/modulos/pesquisa/PesquisaIntegracao.php
@@ -9,7 +9,7 @@ class PesquisaIntegracao extends SeiIntegracao {
 	
 	public function getVersao()
 	{
-		return '4.2.3';
+		return '4.2.5';
 	}
