+++ sei/web/modulos/pesquisa/md_pesq_processo_exibir.php
@@ -12,6 +12,7 @@ try {
     session_start();
 
 	SessaoSEIExterna::getInstance()->validarSessao();
+	PaginaSEIExterna::getInstance()->setTipoPagina(PaginaSEIExterna::$TIPO_PAGINA_SEM_MENU);
  
 //	InfraDebug::getInstance()->setBolLigado(false);
 //	InfraDebug::getInstance()->setBolDebugInfra(true);
@@ -39,7 +40,7 @@ try {
 	$txtDescricaoProcessoAcessoRestrito = $arrParametroPesquisaDTO[MdPesqParametroPesquisaRN::$TA_DESCRICAO_PROCEDIMENTO_ACESSO_RESTRITO];
 	$dtaCortePesquisa = (new MdPesqParametroPesquisaRN())->existeDataCortePesquisa();
 
-	PaginaSEIExterna::getInstance()->setTipoPagina(PaginaSEIExterna::$TIPO_PAGINA_SEM_MENU);
+	
 
     $dblIdProcedimento = $_GET['id_procedimento'];
      
@@ -721,9 +722,7 @@ table caption {
     background-image: url("imagens/bg_barra_sistema.jpg");
     color: white;
 }
-    div {
-    margin: 0 auto 0 auto;
-    }
+
 .modal-body {padding: 2px 16px;}
 
 .modal-footer {
@@ -731,6 +730,19 @@ table caption {
     background-color: #5cb85c;
     color: white;
 }
+
+div:has(> #lnkInfraMenuSistema) {
+    display: none !important;
+}
+
+div:has(> #lnkUsuarioSistema) {
+    display: none !important;
+}
+
+div:has(> #lnkInfraSairSistema) {
+    display: none !important;
+}
+
 <? } ?>
 
 span.retiraAncoraPadraoAzul{font-size: .875rem;}
@@ -740,95 +752,6 @@ PaginaSEIExterna::getInstance()->fecharStyle();
 PaginaSEIExterna::getInstance()->montarJavaScript();
 CaptchaSEI::getInstance()->montarJavascript();
 PaginaSEIExterna::getInstance()->abrirJavaScript();
-?>
-
-function inicializar(){
-
-  <?if ($bolGeracaoOK){?>
-    
-  	window.open('<?=SessaoSEI::getInstance()->assinarLink('md_pesq_processo_exibe_arquivo.php?'.MdPesqCriptografia::criptografa('acao_externa=usuario_externo_exibir_arquivo&acao_origem_externa=protocolo_pesquisar&id_orgao_acesso_externo='.$_GET['id_orgao_acesso_externo'].'&nome_arquivo='.$objAnexoDTO->getStrNome().'&nome_download=SEI-'.$objProcedimentoDTO->getStrProtocoloProcedimentoFormatado().'.pdf'));?>');
-  	
-  <?}?>
-
-  infraEfeitoTabelas();
-}
-
-
-<?
-if($bolCaptchaGerarPdf){ 
-?>
-
-$(document).unbind("keyup").keyup(function(e){
-	e.preventDefault();
-    var code = e.which;
-    if(code==13){
-    	var modal = document.getElementById('divInfraModal');
-        if(modal.style.display == "block"){
-        	fecharPdfModal();
-        	gerarPdf();
-    		
-    	}
-    }
-});
-
-function gerarPdfModal(){
-
-	if (document.getElementById('hdnInfraItensSelecionados').value==''){
-    	alert('Nenhum documento selecionado.');
-    	return;
-  	}
-
-    document.getElementById('divInfraModal').style.display = "block";
-    document.getElementById('txtInfraCaptcha').focus();
-
-}
-
-function fecharPdfModal(){
-	
-    var modal = document.getElementById('divInfraModal');
-	modal.style.display = "none";
-}
-
-window.onclick = function(event) {
-	var modal = document.getElementById('divInfraModal');
-    if (event.target == modal) {
-        modal.style.display = "none";
-    }
-}
-
-<?
-}
-?>
-
-function gerarPdf() {
-    if (document.getElementById('hdnInfraItensSelecionados').value == ''){
-        alert('Nenhum documento selecionado.');
-        return false;
-    }
-
-<? if($bolCaptchaGerarPdf): ?>
-
-    if (document.getElementById('txtInfraCaptcha').value.length != 6){
-        $('.modal-alert-msg').html('<p class="alert alert-warning">Informe o código de confirmação!</p>');
-        document.getElementById('txtInfraCaptcha').focus();
-        return false;
-    }
-
-    if(document.getElementById('hdnInfraItensSelecionados').value != '' && document.getElementById('txtInfraCaptcha').value.length == 6){
-        fecharPdfModal();
-        infraExibirAviso(false);
-        document.getElementById('hdnFlagGerar').value = '1';
-        document.getElementById('frmProcessoAcessoExternoConsulta').submit();
-    }
-<? else: ?>
-
-    document.getElementById('hdnFlagGerar').value = '1';
-    document.getElementById('frmProcessoAcessoExternoConsulta').submit();
-
-<? endif; ?>
-}
-
-<?
 PaginaSEIExterna::getInstance()->fecharJavaScript();
 PaginaSEIExterna::getInstance()->fecharHead();
 PaginaSEIExterna::getInstance()->abrirBody($strTitulo,'onload="inicializar();"');
@@ -957,6 +880,7 @@ if($bolGeracaoOK){
 ?>
 
 <?
+require_once("md_pesq_processo_exibir_js.php");
 PaginaSEIExterna::getInstance()->fecharBody();
 PaginaSEIExterna::getInstance()->fecharHtml();
 ?>
